# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Atlas API Contract
  version: 1.0.0
  summary: Master API contract for Atlas platform clients and services.
  description: |
    Production-grade OpenAPI 3.1 contract for Atlas.

    This specification is the single source of truth for:
    - Web frontend
    - Backend implementation
    - Mobile clients
    - Admin portal
    - External integrations

    Design guarantees:
    - RESTful, versioned API under /api/v1
    - Consistent response envelope
    - Standardized errors
    - Standardized pagination/filter/sort/search
    - Typed request and response contracts

    Note: This contract defines interfaces only. It intentionally contains no implementation logic.
  contact:
    name: Atlas API Platform Team
    email: api-platform@atlas.local
  license:
    name: Proprietary
...
servers:
  - url: https://api.atlasbank.com/api/v1
    description: Production
  - url: https://staging-api.atlasbank.com/api/v1
    description: Staging
  - url: http://localhost:3001/api/v1
    description: Local development
...
security:
  - bearerAuth: []
  - sessionCookie: []
paths:
...
  /accounts:
    get:
      tags: [Accounts]
      summary: List accounts
      description: |
        Returns customer accounts.

        Supports filtering by status/type and standardized sort/search/pagination.

        Permissions: customer:self
        Rate limit: 1000 requests / hour per user
      operationId: listAccounts
      x-permissions: [customer:self]
      x-rateLimit:
        scope: user
        burst: 1000
        window: 1h
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Order'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/AccountStatus'
        - $ref: '#/components/parameters/AccountType'
      responses:
        '200':
          description: Accounts returned.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Account'
                      meta:
                        $ref: '#/components/schemas/CursorMeta'
    post:
      tags: [Accounts]
      summary: Create account
      description: |
        Creates a new account product for the customer.

        Permissions: customer:self
        Rate limit: 20 requests / day per user
      operationId: createAccount
      x-permissions: [customer:self]
      x-rateLimit:
        scope: user
        burst: 20
        window: 24h
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountCreateRequest'
      responses:
        '201':
          description: Account created.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/Conflict'
  /accounts/{accountId}:
    get:
      tags: [Accounts]
      summary: Get account details
      description: |
        Returns a single account by identifier.

        Permissions: customer:self
        Rate limit: 1200 requests / hour per user
      operationId: getAccountById
      x-permissions: [customer:self]
      x-rateLimit:
        scope: user
        burst: 1200
        window: 1h
      parameters:
        - $ref: '#/components/parameters/AccountId'
      responses:
        '200':
          description: Account returned.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Account'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Accounts]
      summary: Update account settings
      description: |
        Updates mutable account preferences and nicknames.

        Permissions: customer:self
        Rate limit: 60 requests / hour per user
      operationId: patchAccountById
      x-permissions: [customer:self]
      x-rateLimit:
        scope: user
        burst: 60
        window: 1h
      parameters:
        - $ref: '#/components/parameters/AccountId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountPatchRequest'
      responses:
        '200':
          description: Account updated.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Account'
...
  /accounts/{accountId}/statements:
    get:
      tags: [Accounts]
      summary: List account statements
      description: |
        Returns account statement history.

        Supports pagination, sorting, and date filtering.

        Permissions: customer:self
        Rate limit: 300 requests / hour per user
      operationId: listAccountStatements
      x-permissions: [customer:self]
      x-rateLimit:
        scope: user
        burst: 300
        window: 1h
      parameters:
        - $ref: '#/components/parameters/AccountId'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Order'
        - $ref: '#/components/parameters/FromDate'
        - $ref: '#/components/parameters/ToDate'
      responses:
        '200':
          description: Statements returned.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Statement'
                      meta:
                        $ref: '#/components/schemas/OffsetMeta'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token for authenticated customer/admin APIs.
    refreshTokenCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: Refresh token cookie used for token refresh endpoint.
    sessionCookie:
      type: apiKey
      in: cookie
      name: atlas_session
      description: Session cookie for session-based flows.
    oauth2:
      type: oauth2
      description: Future OAuth 2.1 support for external integrations.
      flows:
        authorizationCode:
          authorizationUrl: https://auth.atlasbank.com/oauth2/authorize
          tokenUrl: https://auth.atlasbank.com/oauth2/token
          scopes:
            atlas.read: Read access to Atlas resources.
            atlas.write: Write access to Atlas resources.
  parameters:
    Cursor:
      name: cursor
      in: query
      description: Cursor token for forward pagination.
      schema:
        type: string
    Offset:
      name: offset
      in: query
      description: Offset for offset pagination.
      schema:
        type: integer
        minimum: 0
        default: 0
    Limit:
      name: limit
      in: query
      description: Maximum items to return.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 25
    Sort:
      name: sort
      in: query
      description: Field name used for sorting.
      schema:
        type: string
    Order:
      name: order
      in: query
      description: Sort direction.
      schema:
        type: string
        enum: [asc, desc]
        default: desc
    Search:
      name: search
      in: query
      description: Full-text search term.
      schema:
        type: string
        minLength: 1
        maxLength: 120
    FromDate:
      name: fromDate
      in: query
      description: Inclusive start datetime (ISO-8601).
      schema:
        type: string
        format: date-time
    ToDate:
      name: toDate
      in: query
      description: Inclusive end datetime (ISO-8601).
      schema:
        type: string
        format: date-time
    AccountId:
      name: accountId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    AccountStatus:
      name: status
      in: query
      schema:
        type: string
        enum: [ACTIVE, FROZEN, DORMANT, CLOSED, PENDING]
    AccountType:
      name: type
      in: query
      schema:
        type: string
        enum: [CHECKING, SAVINGS]
...
  responses:
...
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            validation:
              value:
                success: false
                data: null
                meta:
                  traceId: trc_4001
                  timestamp: '2026-07-09T13:22:00Z'
                error:
                  type: validation_error
                  code: VAL_001
                  message: One or more fields failed validation.
                  details:
                    - field: email
                      reason: Invalid email format.
...
    NotFound:
      description: Requested resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: Resource conflict or incompatible state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    ApiEnvelope:
      type: object
      required: [success, data, meta, error]
      properties:
        success:
          type: boolean
        data:
          nullable: true
        meta:
          $ref: '#/components/schemas/ResponseMeta'
        error:
          oneOf:
            - type: 'null'
            - $ref: '#/components/schemas/ErrorObject'
    ErrorEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            success:
              const: false
            data:
              type: 'null'
            error:
              $ref: '#/components/schemas/ErrorObject'
    ResponseMeta:
      type: object
      required: [traceId, timestamp]
      properties:
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
        rateLimit:
          $ref: '#/components/schemas/RateLimitMeta'
    PaginationMeta:
      type: object
      properties:
        mode:
          type: string
          enum: [cursor, offset]
        cursor:
          type: object
          properties:
            next:
              type: string
            previous:
              type: string
            hasMore:
              type: boolean
        offset:
          type: object
          properties:
            total:
              type: integer
            offset:
              type: integer
            limit:
              type: integer
    RateLimitMeta:
      type: object
      properties:
        limit:
          type: integer
        remaining:
          type: integer
        resetAt:
          type: string
          format: date-time
    ErrorObject:
      type: object
      required: [type, code, message]
      properties:
        type:
          type: string
          enum:
            - validation_error
            - authentication_error
            - authorization_error
            - business_rule_error
            - rate_limit_error
            - server_error
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              reason:
                type: string
        traceId:
          type: string
    OffsetMeta:
      allOf:
        - $ref: '#/components/schemas/ResponseMeta'
        - type: object
          properties:
            pagination:
              type: object
              properties:
                mode:
                  const: offset
                offset:
                  type: object
                  required: [total, offset, limit]
                  properties:
                    total:
                      type: integer
                    offset:
                      type: integer
                    limit:
                      type: integer
    CursorMeta:
      allOf:
        - $ref: '#/components/schemas/ResponseMeta'
        - type: object
          properties:
            pagination:
              type: object
              properties:
                mode:
                  const: cursor
                cursor:
                  type: object
                  required: [hasMore]
                  properties:
                    next:
                      type: string
                    previous:
                      type: string
                    hasMore:
                      type: boolean
...
    Account:
      type: object
      required: [id, name, type, status]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [CHECKING, SAVINGS]
        status:
          type: string
        accountNumberMasked:
          type: string
        routingNumber:
          type: string
        createdAt:
          type: string
          format: date-time
...
    AccountCreateRequest:
      type: object
      required: [type, name]
      properties:
        type:
          type: string
          enum: [CHECKING, SAVINGS]
        name:
          type: string
          maxLength: 80
    AccountPatchRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          maxLength: 80
        overdraftEnabled:
          type: boolean
    Statement:
      type: object
      properties:
        id:
          type: string
          format: uuid
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        generatedAt:
          type: string
          format: date-time
        downloadUrl:
          type: string
          format: uri
...